requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.23.6
sentence-transformers==2.7.0
pydantic==2.7.1
numpy
pinecone-client==4.1.1
chromadb==0.5.5
PyPDF2==3.0.1
streamlit==1.28.0